% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_index.R
\name{forget_tidysec_cache}
\alias{forget_tidysec_cache}
\title{Clear tidysec cache of master index files.}
\usage{
forget_tidysec_cache()
}
\description{
\code{forget_tidysec_cache} clears the cache
of master index files in which the links to all filings are
document.
}
