% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_index.R
\name{get_13f}
\alias{get_13f}
\title{Retrieve 13F filings}
\usage{
get_13f(cik, year, amendments = FALSE, link_only = FALSE)
}
\arguments{
\item{cik}{Central Index Key for filing entity, see https://www.sec.gov/edgar/searchedgar/companysearch.html}

\item{year}{Year for which filings should be retrieved.}

\item{amendments}{Include 13F-HR/A when retrieving — only works when \code{link_only = TRUE}. Defaults to FALSE.}

\item{link_only}{Do not retrieve full filing but only link to it.}
}
\description{
\code{get_13f} retrieves and parses a company's 13F filings along with meta information or the links to the filings.
}
\examples{
# Get and parse BlackRock's 2015 13F filings.
cik_blackrock <- "1364742"
get_13f(
  cik = cik_blackrock, year = 2015,
  link_only = FALSE
)
# Get the links to the Swiss National Bank's 2016 to 2019 filings.
cik_snb <- "1582202"
snb <- get_13f(
  cik = cik_snb, year = 2016:2019,
  amendments = TRUE, link_only = TRUE
)
}
