% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13f.R
\name{get_filing}
\alias{get_filing}
\title{Retrieve an SEC filing}
\usage{
get_filing(link_to_filing, filing_type = "13F", clean_col_names = TRUE)
}
\arguments{
\item{link_to_filing}{Link to company filing.}

\item{filing_type}{Indicate filing type. Only 13F possible, will fail otherwise.}

\item{clean_col_names}{Remove camel case from column names.}
}
\description{
Parses a company's filing
}
\examples{
snb_q1_2016_13f <- "https://www.sec.gov/Archives/edgar/data/1582202/000158220216000006/InfoTable_Q12016.xml"
xml <- get_filing(snb_q1_2016_13f)
}
