% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13f.R
\name{get_filings_info}
\alias{get_filings_info}
\title{List all filings submitted by company to the SEC.}
\usage{
get_filings_info(cik, clean_col_names = TRUE)
}
\arguments{
\item{cik}{Central index key of a company.}

\item{clean_col_names}{Remove camel case from column names.}
}
\description{
\code{get_filings_info} searches the EDGAR archive for any
filings a company has submitted over time.
}
\details{
The function returns a tibble with all meta information about all filings,
links to the filing itself and the meta information.
}
\examples{
# Retrieve filings made by the Swiss National Bank
cik_snb <- "0001582202"
get_filings_info(cik = cik_snb, clean_col_names = TRUE)
}
