% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_index.R
\name{get_13f}
\alias{get_13f}
\title{Retrieve a 13F filing.}
\usage{
get_13f(
  cik,
  year,
  amendments = FALSE,
  clean_col_names = TRUE,
  link_only = FALSE
)
}
\arguments{
\item{amendments}{Include 13F-HR/A when retrieving — only works when \code{link_only = TRUE}. Defaults to FALSE.}

\item{clean_col_names}{Remove camel case from column names.}

\item{link_only}{Do not retrieve full filing but only link to it.}

\item{link_to_filing}{Link to company filing.}
}
\description{
\code{get_13f} retrieves and parses a company's 13F filings or the links to the filings.
}
\examples{
# Get and parse Blackrock's 2015 13F filings.
cik_blackrock <- "1364742"
get_13f(cik = cik_blackrock, year = 2015, clean_col_names = TRUE, link_only = FALSE)
# Get the links to the Swiss National Bank's 2016 to 2019 filings.
cik_snb <- "1582202"
get_13f(cik = cik_snb, year = 2016:2019, amendments = TRUE, clean_col_names = TRUE, link_only = TRUE)
}
